# docs/catalog-api.openapi.yaml
openapi: 3.0.3
info:
  title: Catalog API
  version: 1.0.0
  description: >
    Catalog and currency rates are public (`lang` selects translation locale with English fallback).
    Checkout purchase requires Bearer JWT with scope catalog.write and an Idempotency-Key header.
    Catalog unit prices are USD; checkout converts with fixed rates for currencyCode.
    Orders start PENDING; stock is decremented when payment-service calls the payment webhook
    with header X-Payment-Secret (no JWT).
servers:
  - url: http://localhost:8090
paths:
...
  /api/v1/checkout/purchase:
    post:
      summary: Create PENDING order and payment session (stock decremented on webhook)
      security:
        - bearerAuth: []
      parameters:
        - in: header
          name: Idempotency-Key
          required: true
          schema: { type: string, maxLength: 64 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Purchase'
      responses:
        '200':
          description: Tracking number and hosted payment URL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurchaseResponse'
        '400':
          description: Validation, inactive product, or insufficient stock
        '401':
          description: Missing or invalid JWT
  /api/v1/checkout/orders/{trackingNumber}:
    get:
      summary: Order payment status for the authenticated customer (SPA result page)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: trackingNumber
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Order status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderStatusResponse'
        '401':
          description: Missing or invalid JWT
        '404':
          description: Unknown order or not owned by caller
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Purchase:
      type: object
      required: [customer, shippingAddress, billingAddress, orderItems, currencyCode]
      properties:
        customer:
          $ref: '#/components/schemas/CustomerRequest'
        shippingAddress:
          $ref: '#/components/schemas/AddressRequest'
        billingAddress:
          $ref: '#/components/schemas/AddressRequest'
        orderItems:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/OrderItemRequest'
        currencyCode:
          type: string
          enum: [USD, EUR, CAD, BRL, INR, TRY]
    CustomerRequest:
      type: object
      required: [firstName, lastName, email]
      properties:
        firstName: { type: string }
        lastName: { type: string }
        email: { type: string, format: email }
    AddressRequest:
      type: object
      required: [street, city, stateId, countryCode, zipCode]
      properties:
        street: { type: string }
        city: { type: string }
        stateId: { type: integer }
        countryCode: { type: string }
        zipCode: { type: string }
    OrderItemRequest:
      type: object
      required: [productId, quantity]
      properties:
        productId: { type: integer, format: int64 }
        quantity: { type: integer, minimum: 1 }
    PurchaseResponse:
      type: object
      required: [orderTrackingNumber, paymentUrl]
      properties:
        orderTrackingNumber: { type: string }
        paymentUrl: { type: string, format: uri }
    OrderStatusResponse:
      type: object
      required: [orderTrackingNumber, status]
      properties:
        orderTrackingNumber: { type: string }
        status:
          type: string
          enum: [PENDING, PAID, CANCELLED]
...